# fwber-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FWBer API Documentation",
        "description": "Comprehensive API documentation for FWBer - A proximity-based dating platform with real-time features, AI-powered matching, and privacy-first design.",
        "contact": {
            "name": "FWBer Team",
            "email": "support@fwber.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.fwber.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/physical-profile": {
            "get": {
                "tags": [
                    "Physical Profile"
                ],
                "summary": "Get physical profile",
                "description": "Get the authenticated user's physical profile",
                "operationId": "da69e43d1596aef74dddf808107adcc9",
                "responses": {
                    "200": {
                        "description": "Profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "height_cm": {
                                                    "type": "integer",
                                                    "nullable": true
                                                },
                                                "body_type": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "hair_color": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "eye_color": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "skin_tone": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "ethnicity": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "facial_hair": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "tattoos": {
                                                    "type": "boolean",
                                                    "nullable": true
                                                },
                                                "piercings": {
                                                    "type": "boolean",
                                                    "nullable": true
                                                },
                                                "dominant_hand": {
                                                    "type": "string",
                                                    "enum": [
                                                        "left",
                                                        "right",
                                                        "ambi"
                                                    ],
                                                    "nullable": true
                                                },
                                                "fitness_level": {
                                                    "type": "string",
                                                    "enum": [
                                                        "low",
                                                        "average",
                                                        "fit",
                                                        "athletic"
                                                    ],
                                                    "nullable": true
                                                },
                                                "clothing_style": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "avatar_prompt": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "avatar_status": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Physical Profile"
                ],
                "summary": "Upsert physical profile",
                "description": "Create or update the authenticated user's physical profile",
                "operationId": "0f5417b826e2de355484dfa9c7f698f5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "height_cm": {
                                        "type": "integer",
                                        "maximum": 250,
                                        "minimum": 80
                                    },
                                    "body_type": {
                                        "type": "string"
                                    },
                                    "hair_color": {
                                        "type": "string"
                                    },
                                    "eye_color": {
                                        "type": "string"
                                    },
                                    "skin_tone": {
                                        "type": "string"
                                    },
                                    "ethnicity": {
                                        "type": "string"
                                    },
                                    "facial_hair": {
                                        "type": "string"
                                    },
                                    "tattoos": {
                                        "type": "boolean"
                                    },
                                    "piercings": {
                                        "type": "boolean"
                                    },
                                    "dominant_hand": {
                                        "type": "string",
                                        "enum": [
                                            "left",
                                            "right",
                                            "ambi"
                                        ]
                                    },
                                    "fitness_level": {
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "average",
                                            "fit",
                                            "athletic"
                                        ]
                                    },
                                    "clothing_style": {
                                        "type": "string"
                                    },
                                    "avatar_prompt": {
                                        "type": "string",
                                        "maxLength": 500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Upserted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/physical-profile/avatar/request": {
            "post": {
                "tags": [
                    "Physical Profile"
                ],
                "summary": "Request avatar generation",
                "description": "Request avatar generation using the saved avatar_prompt",
                "operationId": "92194b031a2d58df173881d60ed56ea5",
                "responses": {
                    "200": {
                        "description": "Requested",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "avatar_status": {
                                                    "type": "string",
                                                    "example": "requested"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Set avatar_prompt first"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}